+++ packages/sdk/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to casper-client-sdk.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.18]
+
+### Added
+
+- Partial support for the Contract object under StoredValue.
+
+### Fixed
+
+- Deploy's body hash derivation.
+
 ## [1.0.17]
 
 ### Added
@@ -13,7 +23,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - Default `gasPrice` changed from `10` to `1`.
-- Casper balances checks return `BigNumber` now. 
+- Casper balances checks return `BigNumber` now.
+
 ## [1.0.15]
 
 ### Added